% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_validation.R
\name{classValidation}
\alias{classValidation}
\title{Validation of binary classification rule}
\usage{
classValidation(model, data, type, scope = NULL,
  validation_method = "bootstrap", B = NULL, B_mat = NULL,
  test_set_method = ifelse(validation_method == "bootstrap", "all",
  "excluded"), show_selected_coef = F, criteria = c(Brier, c_index, calib),
  roc = F, fpr = seq(0, 1, length = 101), ...)
}
\arguments{
\item{model}{Model's formula}

\item{data}{Data to fit and assess model performance}

\item{type}{Type of statistical model}

\item{scope}{Scope in stepwise variable selection}

\item{validation_method}{"cv" (cross-validation) or "bootstrap"}

\item{B}{number of bootstrap samples / B-fold cross-validation}

\item{B_mat}{Data matrix for cross-validation}

\item{test_set_method}{How is the test set defined (either "all" -> use original dataset as test data (default for bootstrap), or "excluded" -> use patients excluded from cv/bootstrap sample as test data)}

\item{show_selected_coef}{If TRUE, show the selected coefficients (only for glm's with model selection)}

\item{criteria}{}

\item{roc}{If TRUE, true positive rates corresponding to false positive rates given in argument fpr are also evaluated and cv/bootstrap corrected}

\item{fpr}{}

\item{...}{Other parameters for fit_method and fit_response}
}
\value{
List of validation output
}
\description{
Validation of binary classification rule
}

