% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_response_methods.R
\name{fit.response}
\alias{fit.response}
\alias{fit.response.cox}
\alias{fit.response.coxtv}
\alias{fit.response.gam}
\alias{fit.response.gbm}
\alias{fit.response.glm}
\alias{fit.response.glmnet1}
\alias{fit.response.glmnet2}
\alias{fit.response.lm}
\alias{fit.response.penalized}
\alias{fit.response.randomForest}
\alias{fit.response.rpart}
\alias{fit.response.scad}
\title{Methods to get response from fitted models}
\usage{
fit.response(fitModel, data, type = c("lm", "glm", "glm.step.AIC",
  "glm.step.BIC", "glm.best.AIC", "glm.best.BIC", "gam", "glmnet1", "glmnet2",
  "adaptiveLASSO", "SCAD", "penalized", "stability", "cox", "cox.step.AIC",
  "cox.step.BIC", "cox.best.AIC", "cox.best.BIC", "glmnet.cox1", "glmnet.cox2",
  "rpart1", "rpart2", "randomForest", "gbm", "gbm.cox", "coxtv",
  "coxtv.step.AIC", "coxtv.step.BIC"), ...)

fit.response.lm(fitModel, data, ...)

fit.response.glm(fitModel, data, ...)

fit.response.cox(fitModel, data, ...)

fit.response.coxtv(fitModel, data, restype = c("risk", "lp"), id,
  save.data = TRUE, t.hor, ...)

fit.response.glmnet1(fitModel, data, ...)

fit.response.glmnet2(fitModel, data, ...)

fit.response.penalized(fitModel, data, ...)

fit.response.scad(fitModel, data, ...)

fit.response.gam(fitModel, data, ...)

fit.response.rpart(fitModel, data, ...)

fit.response.randomForest(fitModel, data, ...)

fit.response.gbm(fitModel, data, ...)
}
\arguments{
\item{fitModel}{Fitted model to get response}

\item{data}{Data to get response}

\item{...}{Other arguments}
}
\value{
Vector of response
}
\description{
Methods to get response from fitted models
}
\section{Functions}{
\itemize{
\item \code{fit.response.lm}: Linear regression

\item \code{fit.response.glm}: Logistic regression

\item \code{fit.response.cox}: Cox regression

\item \code{fit.response.coxtv}: Cox regression with time-dependent variable

\item \code{fit.response.glmnet1}: GLM LASSO (using glmnet, choose the minimum lambda)

\item \code{fit.response.glmnet2}: GLM LASSO (using glmnet, choose the minimum lambda within 1 se)

\item \code{fit.response.penalized}: GLM LASSO (using penalized)

\item \code{fit.response.scad}: GLM with SCAD (choose the minimum lambda)

\item \code{fit.response.gam}: GAM

\item \code{fit.response.rpart}: CART

\item \code{fit.response.randomForest}: Random forest

\item \code{fit.response.gbm}: Boosting
}}

