% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_performance_criteria.R
\name{myaccuracy}
\alias{Brier}
\alias{acc}
\alias{auc}
\alias{best.acc}
\alias{best.youden}
\alias{binperform}
\alias{c_index}
\alias{calib}
\alias{concordance}
\alias{dis_slope}
\alias{loglik_binary}
\alias{mse}
\alias{myaccuracy}
\alias{new.accuracy}
\alias{npv}
\alias{ppv}
\alias{sen}
\alias{spe}
\title{Functions to assess model performance}
\usage{
myaccuracy(predicted, y, criteria = mse, ...)

mse(predicted, y)

loglik_binary(predicted, y)

Brier(predicted, y, confint = FALSE)

dis_slope(predicted, y)

c_index(predicted, y, confint = FALSE, ci_collapse = "-",
  ci_parentheses = FALSE)

concordance(x, y)

calib(predicted, y, confint = FALSE)

binperform(predicted, y, cutoff = 0.5)

sen(predicted, y, cutoff)

spe(predicted, y, cutoff)

ppv(predicted, y, cutoff)

npv(predicted, y, cutoff)

acc(predicted, y, cutoff)

auc(predicted, y, alpha = 0.05, digits = 2)

best.youden(predicted, y, by = 0.01, digits = 2, name = "")

best.acc(predicted, y, by = 0.01, digits = 2, name = "")

new.accuracy(predicted, y, cutoff = 0.5)
}
\arguments{
\item{predicted}{Predicted values from fitted models}

\item{y}{Response}
}
\value{
numeric value of model's performance
}
\description{
Functions to assess model performance
}
\section{Functions}{
\itemize{
\item \code{mse}: Mean squared errors

\item \code{loglik_binary}: log-likelihood (binary case)

\item \code{Brier}: Brier score

\item \code{dis_slope}: Discrimination slope

\item \code{c_index}: AUC (with confidence interval)

\item \code{concordance}: Concordance

\item \code{calib}: Calibration slope and intercept (with confidence intervals)

\item \code{binperform}: Other measures of binary test

\item \code{sen}: Sensitivity

\item \code{spe}: Specificity

\item \code{ppv}: Positive predictive value

\item \code{npv}: Negative predictive value

\item \code{acc}: Accuracy rate

\item \code{auc}: AUC

\item \code{best.youden}: Best Youden's index

\item \code{best.acc}: Best accuracy rate

\item \code{new.accuracy}: Accuracy of binary test
}}

