% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mySummary_set.R
\name{mySummary.onevar}
\alias{mySummary.onevar}
\title{Summarise dataset}
\usage{
mySummary.onevar(varname, variable, group = NULL, continuous = NA,
  contSummary = "med.IQR", test = FALSE, digits = 1)
}
\arguments{
\item{continuous}{Independent variables are continuous variables?}

\item{contSummary}{Summary statistics for continuous variables?}

\item{test}{Display p values from statistical test comparing independent variables between sub-group of dependent variable?}

\item{digits}{Number of decimal digit number}

\item{formula}{Specification of dependent and independent variables}

\item{data}{Data to summarise}

\item{pooledGroup}{Pooling all subgroups of dependent variable?}

\item{caption}{Caption of final kable table}

\item{kable}{Create kable table}
}
\value{
a matrix
}
\description{
Summarise dataset
}

